+++ src/network/win32/winsockwrappers.h
@@ -22,7 +22,7 @@ ERR win_socketstate(WSW_SOCKET, char, char);
 ERR WIN_SEND(WSW_SOCKET, const void *, size_t *, int);
 int win_shutdown(WSW_SOCKET, int);
 WSW_SOCKET win_socket_ipv6(void *, char, char, bool &);/*creates IPv6 dual-stack socket */
-ERR WIN_RECEIVE(WSW_SOCKET, void *, size_t, int, int *);
+ERR WIN_RECEIVE(WSW_SOCKET, void *, size_t, int, size_t *);
 void winCloseResolveHandle(void *);
 void win_socket_reference(WSW_SOCKET, void *);
